[sp] move StatsTable component to new file
