+++ mage_ai/frontend/components/datasets/overview/index.tsx
@@ -273,9 +273,9 @@ function DatasetOverview({
                     <Spacing mb={PADDING_UNITS}>
                       <SimpleDataTable
                         columnFlexNumbers={[2, 1, 2 ]}
-                      columnHeaders={[{ label: 'Quality metrics' }]}
-                      rowGroupData={[qualityMetrics]}
-                      warnings={WARNINGS.qualityMetrics}
+                        columnHeaders={[{ label: 'Quality metrics' }]}
+                        rowGroupData={[qualityMetrics]}
+                        warnings={WARNINGS.qualityMetrics}
                       />
                     </Spacing>
                   )}
